+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 1.4.9 (...)
 - set default for refresh token expiry to 180 days (instead of 6 months to sync
   with default of CA certificates)
+- add Let's Connect OAuth client registration for Windows
 
 ## 1.4.8 (2018-01-24)
 - update Norwegian translation